+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 Changes to Calva.
 
 ## [Unreleased]
+- [Add hover to display results for eval as window into output file](https://github.com/BetterThanTomorrow/calva/issues/693)
 
 ## [2.0.112] - 2020-07-30
 - Fix: [Don't open output window until connect starts](https://github.com/BetterThanTomorrow/calva/issues/707)